change file render.js to index.js, delete useEffect
