--- src/utils/useEffect.js
@@ -1,9 +1,0 @@
-const effects = [];
-const useEffect = (cb) => {
-    effects.push({
-        cb,
-    });
-};
-
-export default useEffect;
-export { effects };
